+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ## [Unreleased]
+
+## [v3.2.0-beta.4]
+  - 📅 2025-01-08
   - 💄 [@mixte/snippets] 优化 `Lottery` 的 `selectCard` 展示卡片的 z 值计算逻辑
   - 🐞 [@mixte/snippets] 修复 `Lottery` 的 `selectCard` 抽取卡片每次都是底部卡片的问题
 
@@ -70,7 +73,8 @@ CHANGELOG 图标规范
 <hr>
 <br>
 
-[Unreleased]: https://github.com/MoomFE/mixte/compare/v3.2.0-beta.3...HEAD
+[Unreleased]: https://github.com/MoomFE/mixte/compare/v3.2.0-beta.4...HEAD
+[v3.2.0-beta.4]: https://github.com/MoomFE/mixte/releases/tag/v3.2.0-beta.4
 [v3.2.0-beta.3]: https://github.com/MoomFE/mixte/releases/tag/v3.2.0-beta.3
 [v3.2.0-beta.2]: https://github.com/MoomFE/mixte/releases/tag/v3.2.0-beta.2
 [v3.2.0-beta.1]: https://github.com/MoomFE/mixte/releases/tag/v3.2.0-beta.1